requirements.txt
Django==3.0.4
psycopg2-binary==2.8.4
python-dateutil==2.8.1
gunicorn==20.0.4
